requirements.txt
requests
pandas
sqlalchemy
python-dotenv
pandas
lxml
html5lib
beautifulsoup4
matplotlib
scikit-learn
scipy
statsmodels
seaborn